Moved Image-related Test Helpers to TestHelpers
